move functions around and moved import settings into import window
