+++ BlenderExtensions/__init__.py
@@ -2,4 +2,3 @@ from .NMSObjectsPanels import NMSPanels  # noqa
 from .CustomNodes import NMSNodes  # noqa
 from .EntityPanels import NMSEntities  # noqa
 from .MiscFunctions import CompareMatrices, ContinuousCompare  # noqa
-from .SettingsPanels import SettingsPanels  # noqa
