+++ __init__.py
@@ -1,9 +1,9 @@
 import bpy
 from .BlenderExtensions import (NMSNodes, NMSEntities, NMSPanels,
-                                NMSShaderNode, SettingsPanels)
-from .addon_script import NMS_Export_Operator, NMS_Import_Operator
+                                NMSShaderNode)
 
 from .NMSDK import ImportSceneOperator, ImportMeshOperator
+from .NMSDK import NMS_Export_Operator, NMS_Import_Operator
 
 customNodes = NMSNodes()
 
@@ -40,7 +40,6 @@ def register():
     NMSShaderNode.register()
     customNodes.register()
     NMSEntities.register()
-    SettingsPanels.register()
     bpy.utils.register_class(ImportSceneOperator)
     bpy.utils.register_class(ImportMeshOperator)
 
@@ -54,6 +53,5 @@ def unregister():
     NMSShaderNode.unregister()
     customNodes.unregister()
     NMSEntities.unregister()
-    SettingsPanels.unregister()
     bpy.utils.unregister_class(ImportSceneOperator)
     bpy.utils.unregister_class(ImportMeshOperator)
